jb-gw: consolidate welcome and recent screens
