+++ components/ide/jetbrains/gateway-plugin/src/main/kotlin/io/gitpod/jetbrains/gateway/GitpodWorkspacesView.kt
@@ -1,0 +1,311 @@
+// Copyright (c) 2022 Gitpod GmbH. All rights reserved.
+// Licensed under the GNU Affero General Public License (AGPL).
+// See License-AGPL.txt in the project root for license information.
+
+package io.gitpod.jetbrains.gateway
+
+import com.intellij.icons.AllIcons
+import com.intellij.ide.BrowserUtil
+import com.intellij.openapi.CompositeDisposable
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.components.service
+import com.intellij.openapi.diagnostic.thisLogger
+import com.intellij.openapi.project.DumbAwareAction
+import com.intellij.openapi.wm.impl.welcomeScreen.WelcomeScreenUIManager
+import com.intellij.remoteDev.util.onTerminationOrNow
+import com.intellij.ui.components.JBScrollPane
+import com.intellij.ui.dsl.builder.BottomGap
+import com.intellij.ui.dsl.builder.RightGap
+import com.intellij.ui.dsl.builder.TopGap
+import com.intellij.ui.dsl.builder.panel
+import com.intellij.ui.dsl.gridLayout.HorizontalAlign
+import com.intellij.ui.dsl.gridLayout.VerticalAlign
+import com.intellij.ui.layout.ComponentPredicate
+import com.intellij.ui.layout.not
+import com.intellij.util.ui.JBFont
+import com.jetbrains.gateway.api.GatewayUI
+import com.jetbrains.rd.util.lifetime.Lifetime
+import com.jetbrains.rd.util.lifetime.LifetimeDefinition
+import com.jetbrains.rd.util.lifetime.isAlive
+import com.jetbrains.rd.util.lifetime.isNotAlive
+import io.gitpod.gitpodprotocol.api.entities.GetWorkspacesOptions
+import io.gitpod.gitpodprotocol.api.entities.WorkspaceInstance
+import io.gitpod.jetbrains.auth.GitpodAuthService
+import io.gitpod.jetbrains.icons.GitpodIcons
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.channels.actor
+import kotlinx.coroutines.ensureActive
+import kotlinx.coroutines.future.await
+import kotlinx.coroutines.launch
+import java.time.OffsetDateTime
+import javax.swing.text.SimpleAttributeSet
+import javax.swing.text.StyleConstants
+import javax.swing.text.StyledDocument
+
+class GitpodWorkspacesView(
+    val lifetime: Lifetime
+) {
+
+    private val settings = service<GitpodSettingsState>()
+
+    private val loggedIn = object : ComponentPredicate() {
+        override fun addListener(listener: (Boolean) -> Unit) {
+            val toDispose = CompositeDisposable()
+            toDispose.add(settings.addListener { listener(invoke()) })
+            toDispose.add(GitpodAuthService.addListener { listener(invoke()) })
+            lifetime.onTerminationOrNow { toDispose.dispose() }
+        }
+
+        override fun invoke(): Boolean {
+            return GitpodAuthService.hasAccessToken(settings.gitpodHost)
+        }
+    }
+
+    private val startWorkspaceView = GitpodStartWorkspaceView(lifetime)
+
+    private lateinit var workspacesPane: JBScrollPane
+    val component = panel {
+        indent {
+            row {
+                panel {
+                    verticalAlign(VerticalAlign.CENTER)
+                    for (i in 1..10) {
+                        row {
+                            label("")
+                        }
+                    }
+                    row {
+                        resizableRow()
+                        icon(GitpodIcons.Logo4x)
+                            .horizontalAlign(HorizontalAlign.CENTER)
+                    }
+                    row {
+                        text(
+                            "Spin up fresh, automated dev environments for each task, in the cloud, in seconds.",
+                            35
+                        ).applyToComponent {
+                            val attrs = SimpleAttributeSet()
+                            StyleConstants.setAlignment(attrs, StyleConstants.ALIGN_CENTER)
+                            (document as StyledDocument).setParagraphAttributes(
+                                0,
+                                document.length - 1,
+                                attrs,
+                                false
+                            )
+                        }.horizontalAlign(HorizontalAlign.CENTER)
+                    }
+                    row {
+                        browserLink("Explore Gitpod", "https://www.gitpod.io")
+                            .horizontalAlign(HorizontalAlign.CENTER)
+                    }.bottomGap(BottomGap.MEDIUM)
+                    row {
+                        button("Connect in Browser") {
+                            GlobalScope.launch {
+                                GitpodAuthService.authorize(settings.gitpodHost)
+                            }
+                        }.horizontalAlign(HorizontalAlign.CENTER)
+                    }
+                }
+            }.visibleIf(loggedIn.not())
+
+            rowsRange {
+                row {
+                    icon(GitpodIcons.Logo).gap(RightGap.SMALL)
+                    label("Gitpod").applyToComponent {
+                        this.font = JBFont.h3().asBold()
+                    }
+                    label("").resizableColumn().horizontalAlign(HorizontalAlign.FILL)
+                    actionsButton(object :
+                        DumbAwareAction("Open Dashboard", "Open Dashboard", AllIcons.Nodes.Servlet) {
+                        override fun actionPerformed(e: AnActionEvent) {
+                            BrowserUtil.browse("https://${settings.gitpodHost}")
+                        }
+                    }, object : DumbAwareAction("Documentation", "Documentation", AllIcons.Toolwindows.Documentation) {
+                        override fun actionPerformed(e: AnActionEvent) {
+                            BrowserUtil.browse("https://www.gitpod.io/docs")
+                        }
+                    }, object : DumbAwareAction("Log Out", "Log out", AllIcons.Actions.Exit) {
+                        override fun actionPerformed(e: AnActionEvent) {
+                            GitpodAuthService.setAccessToken(settings.gitpodHost, null)
+                        }
+                    })
+                    cell()
+                }.topGap(TopGap.MEDIUM).bottomGap(BottomGap.SMALL)
+                row {
+                    cell(startWorkspaceView.component)
+                        .horizontalAlign(HorizontalAlign.FILL)
+                }.bottomGap(BottomGap.SMALL)
+                row {
+                    label("Recent Workspaces").bold()
+                    label("").resizableColumn().horizontalAlign(HorizontalAlign.FILL)
+                    actionButton(object :
+                        DumbAwareAction("Refresh", "Refresh recent workspaces", AllIcons.Actions.Refresh) {
+                        override fun actionPerformed(e: AnActionEvent) {
+                            refresh()
+                        }
+                    })
+                    cell()
+                }
+                row {
+                    resizableRow()
+                    workspacesPane = cell(JBScrollPane())
+                        .resizableColumn()
+                        .horizontalAlign(HorizontalAlign.FILL)
+                        .verticalAlign(VerticalAlign.FILL)
+                        .component
+                    cell()
+                }.bottomGap(BottomGap.SMALL)
+            }.visibleIf(loggedIn)
+        }
+    }.apply {
+        this.background = WelcomeScreenUIManager.getMainAssociatedComponentBackground()
+    }
+
+    val refresh = startUpdateLoop(lifetime, workspacesPane)
+
+    init {
+        refresh()
+        loggedIn.addListener { refresh() }
+    }
+
+    private fun startUpdateLoop(lifetime: Lifetime, workspacesPane: JBScrollPane): () -> Unit {
+        val updateJob = Job()
+        lifetime.onTerminationOrNow { updateJob.cancel() }
+
+        val updateActor = GlobalScope.actor<Void?>(updateJob, capacity = Channel.CONFLATED) {
+            var updateLifetime: LifetimeDefinition? = null
+            for (event in channel) {
+                ensureActive()
+                updateLifetime?.terminate()
+                updateLifetime = lifetime.createNested()
+                doUpdate(updateLifetime, workspacesPane);
+            }
+        }
+        lifetime.onTerminationOrNow { updateActor.close() }
+
+        return { updateActor.trySend(null) }
+    }
+
+    private fun doUpdate(updateLifetime: Lifetime, workspacesPane: JBScrollPane) {
+        val gitpodHost = settings.gitpodHost
+        if (!GitpodAuthService.hasAccessToken(gitpodHost)) {
+            ApplicationManager.getApplication().invokeLater {
+                if (updateLifetime.isAlive) {
+                    workspacesPane.viewport.view = panel {
+                        row {
+                            comment("Loading...")
+                        }
+                    }
+                }
+            }
+            return
+        }
+        val job = GlobalScope.launch {
+            val client = service<GitpodConnectionService>().obtainClient(gitpodHost)
+            val workspaces = client.server.getWorkspaces(GetWorkspacesOptions().apply {
+                this.limit = 20
+            }).await()
+            val workspacesMap = workspaces.associateBy { it.workspace.id }.toMutableMap()
+            fun updateView() {
+                val view = panel {
+                    val sortedInfos = workspacesMap.values.toMutableList()
+                        .sortedByDescending {
+                            val creationTime = it.latestInstance?.creationTime ?: it.workspace.creationTime
+                            try {
+                                if (creationTime != null) {
+                                    OffsetDateTime.parse(creationTime)
+                                } else {
+                                    null
+                                }
+                            } catch (e: Throwable) {
+                                thisLogger().error(
+                                    "${gitpodHost}: ${it.workspace.id}: failed to parse creation time",
+                                    e
+                                )
+                                null
+                            }
+                        }
+                    for (info in sortedInfos) {
+                        if (info.latestInstance == null) {
+                            continue;
+                        }
+                        indent {
+                            row {
+                                var canConnect = false
+                                icon(
+                                    if (info.latestInstance.status.phase == "running") {
+                                        canConnect = true
+                                        GitpodIcons.Running
+                                    } else if (info.latestInstance.status.phase == "stopped") {
+                                        if (info.latestInstance.status.conditions.failed.isNullOrBlank()) {
+                                            GitpodIcons.Stopped
+                                        } else {
+                                            GitpodIcons.Failed
+                                        }
+                                    } else if (info.latestInstance.status.phase == "interrupted") {
+                                        GitpodIcons.Failed
+                                    } else if (info.latestInstance.status.phase == "unknown") {
+                                        GitpodIcons.Failed
+                                    } else {
+                                        canConnect = true
+                                        GitpodIcons.Starting
+                                    }
+                                ).gap(RightGap.SMALL)
+                                panel {
+                                    row {
+                                        browserLink(info.workspace.id, info.latestInstance.ideUrl)
+                                    }.rowComment("<a href='${info.workspace.context.normalizedContextURL}'>${info.workspace.context.normalizedContextURL}</a>")
+                                }
+                                label("").resizableColumn().horizontalAlign(HorizontalAlign.FILL)
+                                button("Connect") {
+                                    if (!canConnect) {
+                                        BrowserUtil.browse(info.latestInstance.ideUrl)
+                                    } else {
+                                        GatewayUI.getInstance().connect(
+                                            mapOf(
+                                                "gitpodHost" to gitpodHost,
+                                                "workspaceId" to info.workspace.id
+                                            )
+                                        )
+                                    }
+                                }
+                                cell()
+                            }
+                        }
+                    }
+                }
+                ApplicationManager.getApplication().invokeLater {
+                    if (updateLifetime.isAlive) {
+                        workspacesPane.viewport.view = view
+                    }
+                }
+            }
+            updateView()
+            val updates = client.listenToWorkspace(updateLifetime, "*")
+            for (update in updates) {
+                if (updateLifetime.isNotAlive) {
+                    return@launch
+                }
+                var info = workspacesMap[update.workspaceId]
+                if (info == null) {
+                    try {
+                        info = client.syncWorkspace(update.workspaceId)
+                    } catch (t: Throwable) {
+                        thisLogger().error("${gitpodHost}: ${update.workspaceId}: failed to sync", t)
+                        continue
+                    }
+                    workspacesMap[update.workspaceId] = info
+                } else if (WorkspaceInstance.isUpToDate(info.latestInstance, update)) {
+                    continue
+                } else {
+                    info.latestInstance = update
+                }
+                updateView()
+            }
+        }
+        updateLifetime.onTerminationOrNow { job.cancel() }
+    }
+}
